\name{CBML-package}
\alias{CBML-package}
\alias{CBML}
\docType{package}
\title{
BML Simulation
}
\description{
A package to implement and analyze the BML simulation process 
}
\details{
\tabular{ll}{
Package: \tab CBML\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-14\cr
License: \tab MIT\cr
}
Three functions are exported: createBMLGrid(), moveCars(), 
plot.BMLGrid().
}
\author{
Xu Han
}
